requirements_dev.txt
pytest>=2.9.1
pytest-cov>=2.2.1
pytest-django>=2.8.0
pylama
pylama-pylint
mccabe
selenium>=2.52.0
django-debug-toolbar

